eliminate differences

+++ src/cradle/contenthandler.tsx
@@ -17,6 +17,8 @@ export default class ContentHandler {
 
       this.cradleParameters = cradleParameters
 
+      // this.internalCallbacksRef = cradleParameters.internalCallbacksRef
+
    }
 
    public content = {
@@ -38,6 +40,8 @@ export default class ContentHandler {
        current:0
     }
     private instanceIdMap = new Map()
+
+    // private internalCallbacksRef
 
     // Two public methods - setCradleContent and updateCradleContent
 
@@ -82,32 +86,16 @@ export default class ContentHandler {
             gap, 
             padding, 
             cellHeight,
-            cellWidth,
-            cache,
-            listsize,
+            cellWidth
         } = cradleInheritedProperties
-
-        const {crosscount} = cradleInternalProperties
-
-        let workingAxisReferenceIndex = Math.min(requestedAxisReferenceIndex,listsize - 1)
-        workingAxisReferenceIndex -= (workingAxisReferenceIndex % crosscount)
-
-        // console.log('cradleState in setCradleContent; workingAxisReferenceIndex',
-        //     cradleState, workingAxisReferenceIndex)
 
         if ((cradleState == 'doreposition') || cradleState == 'reconfigure')  {
 
-            targetAxisPixelOffset = 
-                (workingAxisReferenceIndex == 0)?
-                    padding:
-                    gap // default
+            targetAxisPixelOffset = gap // default
 
         }
 
-        // console.log('cradleState in setCradleContent; workingAxisReferenceIndex, targetAxisPixelOffset',
-        //     cradleState, workingAxisReferenceIndex, targetAxisPixelOffset)
-        
-        const workingContentList = []
+        const localContentList = []
         const cradleContent = this.content
 
         // ----------------------[ 2. get content requirements ]----------------------
@@ -148,7 +136,7 @@ export default class ContentHandler {
         // ----------------------[ 3. get and config content ]----------------------
         
         // returns content constrained by cradleRowcount
-        const [newcontentlist,deleteditems] = getCellShellComponentList({
+        const [childlist,deleteditems] = getCellShellComponentList({
 
             cradleInheritedProperties,
             // cradleInternalProperties,
@@ -156,22 +144,21 @@ export default class ContentHandler {
             cradleReferenceIndex:targetCradleReferenceIndex,
             listStartChangeCount:0,
             listEndChangeCount:cradleContentCount,
-            workingContentList,
+            localContentList,
+            // callbacks:this.internalCallbacksRef.current,
             instanceIdCounterRef:this.instanceIdCounterRef,
         })
 
-        if (deleteditems.length && (cache == 'cradle')) {
-            deletePortals(cacheHandler, deleteditems)
-        }
+        if (deleteditems.length) deletePortals(cacheHandler, deleteditems)
 
         const [headcontentlist, tailcontentlist] = allocateContentList({
 
-            contentlist:newcontentlist,
+            contentlist:childlist,
             axisReferenceIndex:targetAxisReferenceIndex,
     
         })
 
-        cradleContent.cradleModelComponents = newcontentlist
+        cradleContent.cradleModelComponents = childlist
         cradleContent.headModelComponents = headcontentlist
         cradleContent.tailModelComponents = tailcontentlist
 
@@ -192,7 +179,6 @@ export default class ContentHandler {
 
         cradlePositionData.blockScrollPos = scrollblockPixelOffset + scrollPosAdjustment
 
-        console.log('setting SCROLLPOS in setCradleContent', cradlePositionData.blockScrollPos)
         viewportElement[cradlePositionData.blockScrollProperty] =
             cradlePositionData.blockScrollPos
 
@@ -251,15 +237,12 @@ export default class ContentHandler {
         } = this.cradleParameters.handlersRef.current
 
         // scroll data
-        const { scrollData } = scrollHandler
+        const scrollData = scrollHandler.scrollData 
 
         const scrollPos = scrollData.currentupdate
 
-        console.log('updateCradleContent with scrollPos, blockScrollPos, source', 
-            scrollPos, scaffoldHandler.cradlePositionData.blockScrollPos, source)
-
         // first abandon option/3; nothing to do
-        if ( scrollPos < 0) { // for Safari, FF elastic bounce at top of scroll
+        if ( scrollPos < 0) { // for Safari elastic bounce at top of scroll
 
             return
 
@@ -274,10 +257,7 @@ export default class ContentHandler {
         // --------------------[ 2. get shift instruction ]-----------------------
 
         const cradleInheritedProperties = this.cradleParameters.cradleInheritedPropertiesRef.current
-        const { 
-            orientation, 
-            cache,
-        } = cradleInheritedProperties
+        const orientation = cradleInheritedProperties.orientation
 
         // -1 is move a row up to the head, +1 is move a row down to the tail, 0 is no shift
         const shiftinstruction = getShiftInstruction({
@@ -343,24 +323,20 @@ export default class ContentHandler {
                 cradleInheritedProperties,
                 // cradleInternalProperties,
                 cradleContentCount,
-                workingContentList:modelcontentlist,
+                localContentList:modelcontentlist,
                 listStartChangeCount,
                 listEndChangeCount,
                 cradleReferenceIndex:oldCradleReferenceIndex,
+                // callbacks:this.internalCallbacksRef.current,
                 instanceIdCounterRef:this.instanceIdCounterRef,
             })
-
         } else {
 
             updatedContentList = modelcontentlist
 
         }
 
-        if (deletedContentItems.length && (cache == 'cradle')) {
-
-            deletePortals(cacheHandler, deletedContentItems)
-
-        }
+        if (deletedContentItems.length) deletePortals(cacheHandler, deletedContentItems)
 
         // ----------------------------------[ 5. allocate cradle content ]--------------------------
 
@@ -390,9 +366,6 @@ export default class ContentHandler {
                 headcontent.length?
                     cradleInheritedProperties.gap + 'px':
                     0
-            viewportElement.scrollTop = scrollPos
-
-            console.log('topPos = scrollPos + axisPixelOffset, viewportElement.scrollTop',topPos,'=',scrollPos,'+',axisPixelOffset, viewportElement.scrollTop)
 
         } else { // 'horizontal'
